+++ qcengine/tests/test_procedures.py
@@ -29,7 +29,7 @@ def input_data():
     ],
 )
 def test_geometric_psi4(input_data, optimizer, ncores, schema_versions, request):
-    models, _ = schema_versions
+    models, retver, _ = schema_versions
 
     input_data["initial_molecule"] = models.Molecule(**qcng.get_molecule("hydrogen", return_dict=True))
     input_data["input_specification"]["model"] = {"method": "HF", "basis": "sto-3g"}
@@ -43,7 +43,9 @@ def test_geometric_psi4(input_data, optimizer, ncores, schema_versions, request)
     }
 
     input_data = checkver_and_convert(input_data, request.node.name, "pre")
-    ret = qcng.compute_procedure(input_data, optimizer, raise_error=True, task_config=task_config)
+    ret = qcng.compute_procedure(
+        input_data, optimizer, raise_error=True, task_config=task_config, return_version=retver
+    )
     ret = checkver_and_convert(ret, request.node.name, "post")
 
     assert 10 > len(ret.trajectory) > 1
@@ -67,7 +69,7 @@ def test_geometric_psi4(input_data, optimizer, ncores, schema_versions, request)
 @using("psi4")
 @using("geometric")
 def test_geometric_local_options(input_data, schema_versions, request):
-    models, _ = schema_versions
+    models, retver, _ = schema_versions
 
     input_data["initial_molecule"] = models.Molecule(**qcng.get_molecule("hydrogen", return_dict=True))
     input_data["input_specification"]["model"] = {"method": "HF", "basis": "sto-3g"}
@@ -77,7 +79,9 @@ def test_geometric_local_options(input_data, schema_versions, request):
 
     # Set some extremely large number to test
     input_data = checkver_and_convert(input_data, request.node.name, "pre")
-    ret = qcng.compute_procedure(input_data, "geometric", raise_error=True, task_config={"memory": "5000"})
+    ret = qcng.compute_procedure(
+        input_data, "geometric", raise_error=True, task_config={"memory": "5000"}, return_version=retver
+    )
     ret = checkver_and_convert(ret, request.node.name, "post")
 
     assert pytest.approx(ret.trajectory[0].provenance.memory, 1) == 4900
@@ -90,7 +94,7 @@ def test_geometric_local_options(input_data, schema_versions, request):
 @using("rdkit")
 @using("geometric")
 def test_geometric_stdout(input_data, schema_versions, request):
-    models, _ = schema_versions
+    models, retver, _ = schema_versions
 
     input_data["initial_molecule"] = models.Molecule(**qcng.get_molecule("water", return_dict=True))
     input_data["input_specification"]["model"] = {"method": "UFF", "basis": ""}
@@ -99,7 +103,7 @@ def test_geometric_stdout(input_data, schema_versions, request):
     input_data = models.OptimizationInput(**input_data)
 
     input_data = checkver_and_convert(input_data, request.node.name, "pre")
-    ret = qcng.compute_procedure(input_data, "geometric", raise_error=True)
+    ret = qcng.compute_procedure(input_data, "geometric", raise_error=True, return_version=retver)
     ret = checkver_and_convert(ret, request.node.name, "post")
 
     assert ret.success is True
@@ -109,7 +113,7 @@ def test_geometric_stdout(input_data, schema_versions, request):
 @using("psi4")
 @using("berny")
 def test_berny_stdout(input_data, schema_versions, request):
-    models, _ = schema_versions
+    models, retver, _ = schema_versions
 
     input_data["initial_molecule"] = models.Molecule(**qcng.get_molecule("water", return_dict=True))
     input_data["input_specification"]["model"] = {"method": "HF", "basis": "sto-3g"}
@@ -118,7 +122,7 @@ def test_berny_stdout(input_data, schema_versions, request):
     input_data = models.OptimizationInput(**input_data)
 
     input_data = checkver_and_convert(input_data, request.node.name, "pre")
-    ret = qcng.compute_procedure(input_data, "berny", raise_error=True)
+    ret = qcng.compute_procedure(input_data, "berny", raise_error=True, return_version=retver)
     ret = checkver_and_convert(ret, request.node.name, "post")
 
     assert ret.success is True
@@ -128,7 +132,7 @@ def test_berny_stdout(input_data, schema_versions, request):
 @using("psi4")
 @using("berny")
 def test_berny_failed_gradient_computation(input_data, schema_versions, request):
-    models, _ = schema_versions
+    models, retver, _ = schema_versions
 
     input_data["initial_molecule"] = models.Molecule(**qcng.get_molecule("water", return_dict=True))
     input_data["input_specification"]["model"] = {"method": "HF", "basis": "sto-3g"}
@@ -138,7 +142,7 @@ def test_berny_failed_gradient_computation(input_data, schema_versions, request)
     input_data = models.OptimizationInput(**input_data)
 
     input_data = checkver_and_convert(input_data, request.node.name, "pre")
-    ret = qcng.compute_procedure(input_data, "berny", raise_error=False)
+    ret = qcng.compute_procedure(input_data, "berny", raise_error=False, return_version=retver)
     ret = checkver_and_convert(ret, request.node.name, "post", vercheck=False)
 
     assert isinstance(ret, (qcel.models.v1.FailedOperation, qcel.models.v2.FailedOperation))
@@ -149,7 +153,7 @@ def test_berny_failed_gradient_computation(input_data, schema_versions, request)
 @using("geometric")
 @using("rdkit")
 def test_geometric_rdkit_error(input_data, schema_versions, request):
-    models, _ = schema_versions
+    models, retver, _ = schema_versions
 
     input_data["initial_molecule"] = models.Molecule(**qcng.get_molecule("water", return_dict=True)).copy(
         exclude={"connectivity_"}
@@ -160,7 +164,7 @@ def test_geometric_rdkit_error(input_data, schema_versions, request):
     input_data = models.OptimizationInput(**input_data)
 
     input_data = checkver_and_convert(input_data, request.node.name, "pre")
-    ret = qcng.compute_procedure(input_data, "geometric")
+    ret = qcng.compute_procedure(input_data, "geometric", return_version=retver)
     ret = checkver_and_convert(ret, request.node.name, "post", vercheck=False)
 
     assert ret.success is False
@@ -170,7 +174,7 @@ def test_geometric_rdkit_error(input_data, schema_versions, request):
 @using("rdkit")
 @using("geometric")
 def test_optimization_protocols(input_data, schema_versions, request):
-    models, _ = schema_versions
+    models, retver, _ = schema_versions
 
     input_data["initial_molecule"] = models.Molecule(**qcng.get_molecule("water", return_dict=True))
     input_data["input_specification"]["model"] = {"method": "UFF"}
@@ -180,7 +184,7 @@ def test_optimization_protocols(input_data, schema_versions, request):
     input_data = models.OptimizationInput(**input_data)
 
     input_data = checkver_and_convert(input_data, request.node.name, "pre")
-    ret = qcng.compute_procedure(input_data, "geometric", raise_error=True)
+    ret = qcng.compute_procedure(input_data, "geometric", raise_error=True, return_version=retver)
     ret = checkver_and_convert(ret, request.node.name, "post")
 
     assert ret.success, ret.error.error_message
@@ -191,7 +195,7 @@ def test_optimization_protocols(input_data, schema_versions, request):
 
 @using("geometric")
 def test_geometric_retries(failure_engine, input_data, schema_versions, request):
-    models, _ = schema_versions
+    models, retver, _ = schema_versions
 
     failure_engine.iter_modes = ["random_error", "pass", "random_error", "random_error", "pass"]  # Iter 1  # Iter 2
     failure_engine.iter_modes.extend(["pass"] * 20)
@@ -207,7 +211,9 @@ def test_geometric_retries(failure_engine, input_data, schema_versions, request)
     input_data = models.OptimizationInput(**input_data)
 
     input_data = checkver_and_convert(input_data, request.node.name, "pre")
-    ret = qcng.compute_procedure(input_data, "geometric", task_config={"ncores": 13}, raise_error=True)
+    ret = qcng.compute_procedure(
+        input_data, "geometric", task_config={"ncores": 13}, raise_error=True, return_version=retver
+    )
     ret = checkver_and_convert(ret, request.node.name, "post")
 
     assert ret.success is True
@@ -220,7 +226,9 @@ def test_geometric_retries(failure_engine, input_data, schema_versions, request)
     # Ensure we still fail
     failure_engine.iter_modes = ["random_error", "pass", "random_error", "random_error", "pass"]  # Iter 1  # Iter 2
 
-    ret = qcng.compute_procedure(input_data, "geometric", task_config={"ncores": 13, "retries": 1})
+    ret = qcng.compute_procedure(
+        input_data, "geometric", task_config={"ncores": 13, "retries": 1}, return_version=retver
+    )
     ret = checkver_and_convert(ret, request.node.name, "post", vercheck=False)
 
     assert ret.success is False
@@ -286,7 +294,7 @@ def test_geometric_retries(failure_engine, input_data, schema_versions, request)
     ],
 )
 def test_geometric_generic(input_data, program, model, bench, schema_versions, request):
-    models, _ = schema_versions
+    models, retver, _ = schema_versions
 
     input_data["initial_molecule"] = models.Molecule(**qcng.get_molecule("water", return_dict=True))
     input_data["input_specification"]["model"] = model
@@ -296,7 +304,7 @@ def test_geometric_generic(input_data, program, model, bench, schema_versions, r
     }
 
     input_data = checkver_and_convert(input_data, request.node.name, "pre", cast_dict_as="OptimizationInput")
-    ret = qcng.compute_procedure(input_data, "geometric", raise_error=True)
+    ret = qcng.compute_procedure(input_data, "geometric", raise_error=True, return_version=retver)
     ret = checkver_and_convert(ret, request.node.name, "post")
 
     assert ret.success is True
@@ -316,7 +324,7 @@ def test_geometric_generic(input_data, program, model, bench, schema_versions, r
 @using("nwchem")
 @pytest.mark.parametrize("linopt", [0, 1])
 def test_nwchem_relax(linopt, schema_versions, request):
-    models, _ = schema_versions
+    models, retver, _ = schema_versions
 
     # Make the input file
     input_data = {
@@ -330,7 +338,7 @@ def test_nwchem_relax(linopt, schema_versions, request):
 
     # Run the relaxation
     input_data = checkver_and_convert(input_data, request.node.name, "pre")
-    ret = qcng.compute_procedure(input_data, "nwchemdriver", raise_error=True)
+    ret = qcng.compute_procedure(input_data, "nwchemdriver", raise_error=True, return_version=retver)
     ret = checkver_and_convert(ret, request.node.name, "post")
 
     assert 10 > len(ret.trajectory) > 1
@@ -340,7 +348,7 @@ def test_nwchem_relax(linopt, schema_versions, request):
 
 @using("nwchem")
 def test_nwchem_restart(tmpdir, schema_versions, request):
-    models, _ = schema_versions
+    models, retver, _ = schema_versions
 
     # Make the input file
     input_data = {
@@ -357,12 +365,16 @@ def test_nwchem_restart(tmpdir, schema_versions, request):
     local_opts = {"scratch_messy": True, "scratch_directory": str(tmpdir)}
 
     input_data = checkver_and_convert(input_data, request.node.name, "pre")
-    ret = qcng.compute_procedure(input_data, "nwchemdriver", task_config=local_opts, raise_error=False)
+    ret = qcng.compute_procedure(
+        input_data, "nwchemdriver", task_config=local_opts, raise_error=False, return_version=retver
+    )
     ret = checkver_and_convert(ret, request.node.name, "post", vercheck=False)
     assert not ret.success
 
     # Run it again, which should converge
-    new_ret = qcng.compute_procedure(input_data, "nwchemdriver", task_config=local_opts, raise_error=True)
+    new_ret = qcng.compute_procedure(
+        input_data, "nwchemdriver", task_config=local_opts, raise_error=True, return_version=retver
+    )
     new_ret = checkver_and_convert(new_ret, request.node.name, "post")
     assert new_ret.success
 
@@ -370,7 +382,7 @@ def test_nwchem_restart(tmpdir, schema_versions, request):
 @using("rdkit")
 @using("torsiondrive")
 def test_torsiondrive_generic(schema_versions, request):
-    models, _ = schema_versions
+    models, retver, _ = schema_versions
     TorsionDriveInput = models.TorsionDriveInput
     TDKeywords = models.procedures.TDKeywords
     OptimizationSpecification = models.procedures.OptimizationSpecification
@@ -394,7 +406,7 @@ def test_torsiondrive_generic(schema_versions, request):
     )
 
     input_data = checkver_and_convert(input_data, request.node.name, "pre")
-    ret = qcng.compute_procedure(input_data, "torsiondrive", raise_error=True)
+    ret = qcng.compute_procedure(input_data, "torsiondrive", raise_error=True, return_version=retver)
     ret = checkver_and_convert(ret, request.node.name, "post")
 
     assert ret.error is None
@@ -430,7 +442,7 @@ def test_torsiondrive_generic(schema_versions, request):
     ],
 )
 def test_optimization_mrchem(input_data, optimizer, schema_versions, request):
-    models, _ = schema_versions
+    models, retver, _ = schema_versions
 
     input_data["initial_molecule"] = models.Molecule(**qcng.get_molecule("hydrogen", return_dict=True))
     input_data["input_specification"]["model"] = {"method": "HF"}
@@ -440,7 +452,7 @@ def test_optimization_mrchem(input_data, optimizer, schema_versions, request):
     input_data = models.OptimizationInput(**input_data)
 
     input_data = checkver_and_convert(input_data, request.node.name, "pre")
-    ret = qcng.compute_procedure(input_data, optimizer, raise_error=True)
+    ret = qcng.compute_procedure(input_data, optimizer, raise_error=True, return_version=retver)
     ret = checkver_and_convert(ret, request.node.name, "post")
 
     assert 10 > len(ret.trajectory) > 1